Teacher Lesson Delete

+++ src/components/SelectedLessonList.jsx
@@ -1,4 +1,4 @@
-import React, { useState, memo } from "react";
+import React, { useRef, memo } from "react";
 import axios from "axios";
 
 const SelectedLessonList = ({
@@ -6,10 +6,13 @@ const SelectedLessonList = ({
 	selectedDate,
 	selectedLessons,
 	joinedAcademyList,
-	joinedClassroomList
+	joinedClassroomList,
+	yearMonth
 }) => {
-	// 수업 삭제 후 리렌더링을 위한 flag state
-	const [isDeleteLesson, setIsDeleteLesson] = useState(false);
+	const ulSelectedLessons = useRef();
+	const divSelectedLesson = useRef();		// 하단에서 선택한 lesson의 div
+	const liLessonInfo = useRef();
+	const btnDeleteLesson = useRef();
 
 	/* academyId로 academyName 검색 */
 	const getAcademyName = (academyId) => {
@@ -38,8 +41,10 @@ const SelectedLessonList = ({
 			const endTime = selectedLessons[i].schedule.endTime.slice(0, 5);
 
 			const element = (
-				<div className="div__selected_lesson_list_menu">
-					<li key={selectedLessons[i].id}>
+				// <div className="div__selected_lesson_list_menu" ref={divSelectedLesson}
+				// 	onClick={(e) => console.log(divSelectedLesson.current.firstElementChild.childNodes)}>
+				<div className="div__selected_lesson_list_menu" ref={divSelectedLesson}>
+					<li key={selectedLessons[i].id} ref={liLessonInfo}>
 						{academyName} 학원, {classroomName} 반, {lessonName}{" "}
 						수업, {beginTime} ~ {endTime}
 					</li>
@@ -47,7 +52,8 @@ const SelectedLessonList = ({
 					{/* 수업 일정 수정, 수업 일정 삭제 button */}
 					{userType === "teacher" ? (
 						<div className="div__selected_lesson_list_menu_buttons">
-							<button onClick={onClickBtnEditLesson}>📝</button>
+							<button
+								onClick={onClickBtnEditLesson}>📝</button>
 							<button
 								onClick={(e) =>
 									onClickBtnDeleteLesson(
@@ -55,6 +61,7 @@ const SelectedLessonList = ({
 										e
 									)
 								}
+								ref={btnDeleteLesson}
 							>
 								❌
 							</button>
@@ -74,10 +81,23 @@ const SelectedLessonList = ({
 	/* 수업 일정 수정 button 📝 클릭 */
 	const onClickBtnEditLesson = (e) => {
 		alert("수업 일정 수정");
+		// console.log(divSelectedLesson.current.childNodes[0]);
 	};
 
 	/* 수업 일정 삭제 button ❌ 클릭 */
 	const onClickBtnDeleteLesson = (lessonId, e) => {
+		console.log("onClickBtnDeleteLesson() !!!!!!");
+
+		// console.log(divSelectedLesson.current);
+
+		// const academyName = divSelectedLesson.current.firstElementChild.childNodes[0];
+		// const classroomName = divSelectedLesson.current.firstElementChild.childNodes[2];
+		// const lessonName = divSelectedLesson.current.firstElementChild.childNodes[4];
+		// const beginTime = divSelectedLesson.current.firstElementChild.childNodes[7];
+		// const endTime = divSelectedLesson.current.firstElementChild.childNodes[9];
+
+		// console.log(academyName, classroomName, lessonName, beginTime, endTime);
+
 		const deleteLesson = async (lessonId) => {
 			try {
 				const response = await axios.delete(
@@ -92,17 +112,26 @@ const SelectedLessonList = ({
 			}
 		};
 
-		deleteLesson(lessonId);
-		// setIsDeleteLesson(true);
+		// console.log(academyName, classroomName, lessonName, beginTime, endTime);
 
-		// 컴포넌트 리렌더링
+		let confirmDeleteLesson = window.confirm(
+			// `${academyName} 학원, ${classroomName} 반, ${lessonName} 수업(${beginTime} ~ ${endTime})을 삭제 하시겠습니까?`
+			"수업 삭제???"
+		);
+		if (confirmDeleteLesson) {
+			deleteLesson(lessonId);
+
+			// 페이지 새로고침
+			window.location.replace(`/teacher/main/calendar_year_month/${yearMonth}`);
+			// history.push(`/teacher/main/calendar_year_month/${yearMonth}`);
+		}
 	};
 
 	return (
 		<div className="div__selected_lesson_list">
 			{selectedDate !== "" ? <h3>{selectedDate} 수업 일정</h3> : ""}
 
-			<ul className="ul__selected_lesson_list">
+			<ul className="ul__selected_lesson_list" ref={ulSelectedLessons} onClick={(e) => console.log(ulSelectedLessons.current.parentNode.)}>
 				{selectedLessons?.length === 0 ? (
 					<span>수업이 없습니다!</span>
 				) : (
